moving users under areas
