+++ src/app/modules/admin/areas/index.ts
@@ -1,0 +1,18 @@
+import { COMPONENTS as UCMP, SERVICES as USRV, STATE_PROVIDERS as USTATE, EFFECTS as UEFF, ROUTES as UR } from "./users";
+
+export const COMPONENTS = [
+	...UCMP
+]
+
+export const PROVIDERS = [
+	...USTATE,
+	...USRV
+]
+
+export const EFFECTS = [
+	...UEFF
+]
+
+export const ROUTES = [
+	...UR
+]
