+++ src/app/modules/admin/areas/users/state/index.ts
@@ -1,5 +1,10 @@
 import { UsersActions } from "./users.actions";
+import { EFFECTS_MODULE } from "./users.effects";
 
 export const STATE_PROVIDERS = [
 	UsersActions
+]
+
+export const EFFECTS = [
+	EFFECTS_MODULE
 ]
