+++ src/app/modules/admin/areas/users/state/users.reducer.ts
@@ -1,5 +1,5 @@
 import { ServerCollection, UserData, EditedUser } from "./users.model";
-import { tassign } from "./../../../state";
+import { tassign } from "./../../../../../state";
 import { Action } from '@ngrx/store';
 import { UsersActions } from "./users.actions";
